cleanup on analyzers and moving things into a single package
