+++ pkg/skaffold/initializer/kubectl_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package kubectl
+package initializer
 
 import (
 	"testing"
@@ -38,7 +38,7 @@ spec:
 `)
 	filename := tmpDir.Path("deployment.yaml")
 
-	k, err := New([]string{filename})
+	k, err := newKubectlInitializer([]string{filename})
 	if err != nil {
 		t.Fatal("failed to create a pipeline")
 	}
@@ -50,7 +50,7 @@ spec:
 			},
 		},
 	}
-	testutil.CheckDeepEqual(t, expectedConfig, k.GenerateDeployConfig())
+	testutil.CheckDeepEqual(t, expectedConfig, k.deployConfig())
 }
 
 func TestParseImagesFromKubernetesYaml(t *testing.T) {
